+++ src/main/java/com/telran/org/urlshortener/service/UrlBindingServiceImpl.java
@@ -9,8 +9,9 @@ import com.telran.org.urlshortener.exception.IdNotFoundException;
 import com.telran.org.urlshortener.exception.PathPrefixNotAvailableException;
 import com.telran.org.urlshortener.mapper.Converter;
 import com.telran.org.urlshortener.repository.UrlBindingJpaRepository;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.nio.charset.StandardCharsets;
 
@@ -21,25 +22,23 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
-@Transactional
+@RequiredArgsConstructor
+@Slf4j
 public class UrlBindingServiceImpl implements UrlBindingService {
-    private UrlBindingJpaRepository repository;
+    private final UrlBindingJpaRepository repository;
 
-    private UserService service;
-
-    private Converter<UrlBinding, UrlBindingCreateDTO, UrlBindingDTO> converter;
+    private final Converter<UrlBinding, UrlBindingCreateDTO, UrlBindingDTO> converter;
 
     @Override
     public UrlBindingDTO create(UrlBindingCreateDTO dto, String pathPrefix) {
-        User currentUser = service.findById(0);
-        String pathSuffix = Base64.getEncoder()
-                .encodeToString(dto.getOriginalUrl().getBytes(StandardCharsets.UTF_8)).substring(1, 8);
+        User currentUser = new User();
+        String pathSuffix = Base64.getEncoder().withoutPadding()
+                .encodeToString(dto.getOriginalUrl().getBytes(StandardCharsets.UTF_8)).substring(0, 8);
         if (pathPrefix == null || pathPrefix.isEmpty()) {
-            Optional<UrlBinding> noPrefix = repository.findAll().stream()
-                    .filter(ub -> ub.getUId().equals("/" + pathSuffix)).findFirst();
+            Optional<UrlBinding> noPrefix = repository.findByUId("/" + pathSuffix);
             if (noPrefix.isPresent()) {
                 noPrefix.get().setUser(currentUser);
-                return converter.entityToDto(noPrefix.get());
+                return converter.entityToDto(repository.save(noPrefix.get()));
             }
             UrlBinding urlBindingBeforeDB = converter.dtoToEntity(dto);
             urlBindingBeforeDB.setUId("/" + pathSuffix);
@@ -50,26 +49,25 @@ public class UrlBindingServiceImpl implements UrlBindingService {
                 .filter(s -> s.getPathPrefix().equals(pathPrefix))
                 .filter(s -> s.getExpirationDate().isAfter(LocalDate.now())).findFirst();
         if (userActivePrefix.isPresent()) {
-            Optional<UrlBinding> withPrefix = repository.findAll().stream()
-                    .filter(ub -> ub.getUId().equals(pathPrefix + "/" + pathSuffix)).findFirst();
+            Optional<UrlBinding> withPrefix = repository.findByUId(pathPrefix + "/" + pathSuffix);
             if (withPrefix.isPresent()) {
                 withPrefix.get().setUser(currentUser);
                 withPrefix.get().setCount(0L);
-                return converter.entityToDto(withPrefix.get());
+                return converter.entityToDto(repository.save(withPrefix.get()));
             }
             UrlBinding urlBindingBeforeDB = converter.dtoToEntity(dto);
             urlBindingBeforeDB.setUId(pathPrefix + "/" + pathSuffix);
             urlBindingBeforeDB.setUser(currentUser);
             return converter.entityToDto(repository.save(urlBindingBeforeDB));
         }
-        throw new PathPrefixNotAvailableException("Can not create UrlBinding, as Subscription with prefix \""
+        throw new PathPrefixNotAvailableException("Can not create Url binding, as Subscription with prefix \""
                 + pathPrefix + "\" is neither active nor belongs to user.");
     }
 
     @Override
     public List<UrlBindingDTO> findAllByUserId(long userId) {
-        return service.findById(userId).getUrlBindings().stream()
-                .map(ub -> converter.entityToDto(ub)).collect(Collectors.toList());
+        return repository.findByUserId(userId).stream()
+                .map(converter::entityToDto).collect(Collectors.toList());
     }
 
     @Override
@@ -82,13 +80,14 @@ public class UrlBindingServiceImpl implements UrlBindingService {
     @Override
     public UrlBindingDTO reset(long id) {
         UrlBinding urlBindingFromDB = repository.findById(id)
-                .orElseThrow(() -> new IdNotFoundException("UrlBinding with id " + id + " is not found."));
+                .orElseThrow(() -> new IdNotFoundException("Url binding with id " + id + " is not found."));
         urlBindingFromDB.setCount(0L);
-        return converter.entityToDto(urlBindingFromDB);
+        return converter.entityToDto(repository.save(urlBindingFromDB));
     }
 
     @Override
-    public void remove(long id) {
-        repository.deleteById(id);
+    public void delete(long id) {
+        if (repository.existsById(id)) repository.deleteById(id);
+        else throw new IdNotFoundException("Url binding with id " + id + " is not found.");
     }
 }
