+++ src/main/java/com/telran/org/urlshortener/service/UserServiceImpl.java
@@ -8,19 +8,23 @@ import com.telran.org.urlshortener.exception.UserNotFoundException;
 import com.telran.org.urlshortener.mapper.Converter;
 import com.telran.org.urlshortener.model.RoleType;
 import com.telran.org.urlshortener.repository.UserJpaRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
-@Transactional
+@RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
-    private UserJpaRepository repository;
+    private final UserJpaRepository repository;
 
-    private Converter<User, UserCreateDTO, UserDTO> converter;
+    private final PasswordEncoder encoder;
+
+    private final Converter<User, UserCreateDTO, UserDTO> converter;
 
     @Override
     public UserDTO create(UserCreateDTO dto) {
@@ -28,37 +32,42 @@ public class UserServiceImpl implements UserService {
         if (userFromDB.isPresent()) {
             throw new EmailNotUniqueException("User with email " + dto.getEmail() + " already exists.");
         }
-        User userBeforeDB = converter.dtoToEntity(dto);
-        return converter.entityToDto(repository.save(userBeforeDB));
+        dto.setPassword(encoder.encode(dto.getPassword()));
+        return converter.entityToDto(repository.save(converter.dtoToEntity(dto)));
     }
 
     @Override
-    public List<UserDTO> findAll() {
+    @PreAuthorize("hasRole('ADMIN')")
+    public List<UserDTO> findUsers() {
         return repository.findAll().stream()
-                .map(u -> converter.entityToDto(u)).collect(Collectors.toList());
+                .map(converter::entityToDto).collect(Collectors.toList());
     }
 
     @Override
-    public User findById(long id) {//UserDTO
-        return repository.findById(id)
-                //.map(u -> convertor.entityToDto(u))
+    @PreAuthorize("hasRole('ADMIN')")
+    public UserDTO findById(long id) {
+        return repository.findById(id).map(converter::entityToDto)
                 .orElseThrow(() -> new UserNotFoundException("User with id " + id + " is not found."));
     }
 
     @Override
+    @PreAuthorize("hasRole('ADMIN')")
     public UserDTO update(long id, RoleType newRole) {
         User userFromDB = repository.findById(id)
                 .orElseThrow(() -> new UserNotFoundException("User with id " + id + " is not found."));
         userFromDB.setRole(newRole);
-        return converter.entityToDto(userFromDB);
+        return converter.entityToDto(repository.save(userFromDB));
     }
 
     @Override
-    public void remove(long id) {
-        repository.deleteById(id);
+    @PreAuthorize("hasRole('ADMIN')")
+    public void delete(long id) {
+        if (repository.existsById(id)) repository.deleteById(id);
+        else throw new UserNotFoundException("User with id " + id + " is not found.");
     }
 
     @Override
+    @PreAuthorize("hasRole('ADMIN')")
     public UserDTO findByEmail(String email) {
         User userFromDB = repository.findByEmail(email)
                 .orElseThrow(() -> new UserNotFoundException("User with email " + email + " is not found."));
